finished padding funcs

+++ crypto/sha_pad.h
@@ -6,6 +6,6 @@ unsigned char * sha256_pad(unsigned char *, long);
 
 
 /* pad for SHA-{384,512,512/224,512/256} length of output in bytes will be largest multiple of 128 closest to orig len */
-unsigned char * sha512_pad(unsigned char *, long);
+unsigned char * sha512_pad(unsigned char *, long long);
 
 #endif /* __BKSMT_CRYPTO_SHA_PAD_H__ */
